refactor: consolidate AI client utils & centralize env loading
